backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dateutil==2.9.0.post0
openpyxl==3.1.5
requests==2.32.3

